fix RawZipkinTracer when traces.length > 10
I noticed in my testing that using node-tryfer with RawZipkinTracer
directly worked, but using ZipkinTracer with the buffering casued traces
to go missing. After some digging I discovered that
RawZipkinTracer.sendTraces is doing:

    async.forEachLimit(traces, 10, this._sendTrace.bind(this),
                      callback);

The problem is that the _sendTrace method wasn't accepting a callback,
so if there are more than 10 traces they just get stuck in limbo
forever. This patch fixes _sendTrace to use the callback that is sent to
it.